# services/product-service/docs/v1/commerce/service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "COMMERCE SERVICE API",
    "description": "API for COMMERCE, authentication, and authorization",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "CartService"
    },
    {
      "name": "WishlistService"
    }
  ],
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/commerce/carts": {
      "get": {
        "summary": "get all cart item with pagination",
        "operationId": "CartService_FindCartItemsWithPagination",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/commerceFindCartItemsWithPaginationResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "productIds",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "CART"
        ]
      },
      "post": {
        "summary": "add to cart",
        "operationId": "CartService_AddToCart",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/commerceAddToCartResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/commerceAddToCartRequest"
            }
          }
        ],
        "tags": [
          "CART"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/commerce/carts/{id}": {
      "delete": {
        "summary": "delete cart item by provided id",
        "operationId": "CartService_DeleteCartItemById",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/commerceDeleteCartItemByIdResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "CART"
        ]
      }
    },
    "/v1/commerce/wishlists": {
      "get": {
        "summary": "get all wishlist item with pagination",
        "operationId": "WishlistService_FindWishlistItemWithPagination",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/commerceFindWishlistItemWithPaginationResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "productIds",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "WISHLIST"
        ]
      },
      "post": {
        "summary": "create a new wishlist",
        "operationId": "WishlistService_AddToWishlist",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/commerceAddToWishlistResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/commerceAddToWishlistRequest"
            }
          }
        ],
        "tags": [
          "WISHLIST"
        ]
      }
    },
    "/v1/commerce/wishlists/{id}": {
      "delete": {
        "summary": "delete wishlist item with provided id",
        "operationId": "WishlistService_DeleteWishlistItemById",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/commerceDeleteWishlistItemByIdResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "WISHLIST"
        ]
      }
    }
  },
  "definitions": {
    "AddToCartResponseAddToCartResponseData": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "AddToWishlistResponseAddToWishlistResponseData": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "DeleteCartItemByIdResponseDeleteCartItemByIdResponseData": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "DeleteWishlistItemByIdResponseDeleteWishlistItemByIdResponseData": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        }
      }
    },
    "FindCartItemsWithPaginationResponseFindCartItemsWithPaginationResponseData": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/commerceCartItem"
          }
        },
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        },
        "total": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "FindWishlistItemWithPaginationResponseFindWishlistItemWithPaginationResponseData": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/commerceWishlistItem"
          }
        },
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        },
        "total": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "commerceAddToCartRequest": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "string"
        },
        "qty": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "CART ITEM REQUEST DEFINITION"
    },
    "commerceAddToCartResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/AddToCartResponseAddToCartResponseData"
        }
      },
      "title": "CART ITEM RESPONSE DEFINITION"
    },
    "commerceAddToWishlistRequest": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "string"
        }
      },
      "title": "WISHLIST ITEM REQUEST DEFINITION"
    },
    "commerceAddToWishlistResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/AddToWishlistResponseAddToWishlistResponseData"
        }
      },
      "title": "WISHLIST ITEM RESPONSE DEFINITION"
    },
    "commerceCartItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "productId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "qty": {
          "type": "integer",
          "format": "int32"
        },
        "price": {
          "type": "number",
          "format": "double"
        },
        "cratedAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "commerceDeleteCartItemByIdResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/DeleteCartItemByIdResponseDeleteCartItemByIdResponseData"
        }
      }
    },
    "commerceDeleteWishlistItemByIdResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/DeleteWishlistItemByIdResponseDeleteWishlistItemByIdResponseData"
        }
      }
    },
    "commerceFindCartItemsWithPaginationResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/FindCartItemsWithPaginationResponseFindCartItemsWithPaginationResponseData"
        }
      }
    },
    "commerceFindWishlistItemWithPaginationResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/FindWishlistItemWithPaginationResponseFindWishlistItemWithPaginationResponseData"
        }
      }
    },
    "commerceWishlistItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "productId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "cratedAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "description": "Enter the token with the `Bearer ` prefix, e.g. \"Bearer abcde12345\"",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}